+++ internal/runner/strategy/contract.go
@@ -3,5 +3,5 @@ package strategy
 import "os/exec"
 
 type ServerRunner interface {
-	BuildCommand(javaPath string, absServerDir string, ram int) (*exec.Cmd, error)
+	BuildCommand(javaPath string, serverDir string, ram int, customArgs string) (*exec.Cmd, error)
 }